Move friend.jsx to parent folder
