+++ client/src/components/friend.jsx
@@ -2,7 +2,7 @@ import { Flex, Button } from '@chakra-ui/react';
 import React, { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import axios from 'axios';
-import pepe from '../../../public/images/PepeProfile.jpeg';
+import pepe from '../../public/images/PepeProfile.jpeg';
 
 function Friend({ user_id, friend, event_id }) {
   const navigate = useNavigate();
